Prepare for upcoming pytest 7.0 deprecations (part 2)
This follows up #42, we decided that using just `fspath` as the new
parameter that users should use would be confusing because it is too
close to `path` and also we have `fspath` already in some places of
the pytest public API (see pytest-dev/pytest#9283).

This PR updates the hook affected by this new change in pytest-dev/pytest#9363.